+++ Engine/NatronEngine/exprutils_wrapper.h
@@ -1,0 +1,9 @@
+#ifndef SBK_EXPRUTILS_H
+#define SBK_EXPRUTILS_H
+
+#include <shiboken.h>
+
+#include <PyExprUtils.h>
+
+#endif // SBK_EXPRUTILS_H
+