+++ Engine/NatronEngine/natronengine_python.h
@@ -23,6 +23,7 @@ CLANG_DIAG_ON(uninitialized)
 #include <RectD.h>
 #include <RectI.h>
 #include <PyTracker.h>
+#include <PyExprUtils.h>
 #include <PyParameter.h>
 #include <PyRoto.h>
 #include <PyGlobalFunctions.h>
@@ -56,71 +57,72 @@ CLANG_DIAG_ON(uninitialized)
 #include <set>
 
 // Type indices
-#define SBK_NATRON_NAMESPACE_IDX                                     29
-#define SBK_NATRON_NAMESPACE_STANDARDBUTTONENUM_IDX                  40
-#define SBK_QFLAGS_NATRON_NAMESPACE_STANDARDBUTTONENUM__IDX          52
-#define SBK_NATRON_NAMESPACE_IMAGECOMPONENTSENUM_IDX                 33
-#define SBK_NATRON_NAMESPACE_IMAGEBITDEPTHENUM_IDX                   32
-#define SBK_NATRON_NAMESPACE_KEYFRAMETYPEENUM_IDX                    35
-#define SBK_NATRON_NAMESPACE_MERGINGFUNCTIONENUM_IDX                 36
-#define SBK_NATRON_NAMESPACE_VALUECHANGEDREASONENUM_IDX              42
-#define SBK_NATRON_NAMESPACE_ANIMATIONLEVELENUM_IDX                  30
-#define SBK_NATRON_NAMESPACE_ORIENTATIONENUM_IDX                     37
-#define SBK_NATRON_NAMESPACE_DISPLAYCHANNELSENUM_IDX                 31
-#define SBK_NATRON_NAMESPACE_IMAGEPREMULTIPLICATIONENUM_IDX          34
-#define SBK_NATRON_NAMESPACE_STATUSENUM_IDX                          41
-#define SBK_NATRON_NAMESPACE_VIEWERCOMPOSITINGOPERATORENUM_IDX       44
-#define SBK_NATRON_NAMESPACE_PLAYBACKMODEENUM_IDX                    39
-#define SBK_NATRON_NAMESPACE_PIXMAPENUM_IDX                          38
-#define SBK_NATRON_NAMESPACE_VIEWERCOLORSPACEENUM_IDX                43
-#define SBK_RECTD_IDX                                                53
-#define SBK_RECTI_IDX                                                54
+#define SBK_NATRON_NAMESPACE_IDX                                     30
+#define SBK_NATRON_NAMESPACE_STANDARDBUTTONENUM_IDX                  41
+#define SBK_QFLAGS_NATRON_NAMESPACE_STANDARDBUTTONENUM__IDX          53
+#define SBK_NATRON_NAMESPACE_IMAGECOMPONENTSENUM_IDX                 34
+#define SBK_NATRON_NAMESPACE_IMAGEBITDEPTHENUM_IDX                   33
+#define SBK_NATRON_NAMESPACE_KEYFRAMETYPEENUM_IDX                    36
+#define SBK_NATRON_NAMESPACE_MERGINGFUNCTIONENUM_IDX                 37
+#define SBK_NATRON_NAMESPACE_VALUECHANGEDREASONENUM_IDX              43
+#define SBK_NATRON_NAMESPACE_ANIMATIONLEVELENUM_IDX                  31
+#define SBK_NATRON_NAMESPACE_ORIENTATIONENUM_IDX                     38
+#define SBK_NATRON_NAMESPACE_DISPLAYCHANNELSENUM_IDX                 32
+#define SBK_NATRON_NAMESPACE_IMAGEPREMULTIPLICATIONENUM_IDX          35
+#define SBK_NATRON_NAMESPACE_STATUSENUM_IDX                          42
+#define SBK_NATRON_NAMESPACE_VIEWERCOMPOSITINGOPERATORENUM_IDX       45
+#define SBK_NATRON_NAMESPACE_PLAYBACKMODEENUM_IDX                    40
+#define SBK_NATRON_NAMESPACE_PIXMAPENUM_IDX                          39
+#define SBK_NATRON_NAMESPACE_VIEWERCOLORSPACEENUM_IDX                44
+#define SBK_RECTD_IDX                                                54
+#define SBK_RECTI_IDX                                                55
 #define SBK_COLORTUPLE_IDX                                           9
 #define SBK_DOUBLE3DTUPLE_IDX                                        13
 #define SBK_DOUBLE2DTUPLE_IDX                                        11
-#define SBK_INT3DTUPLE_IDX                                           24
-#define SBK_INT2DTUPLE_IDX                                           22
-#define SBK_PARAM_IDX                                                48
-#define SBK_PAGEPARAM_IDX                                            47
-#define SBK_GROUPPARAM_IDX                                           19
-#define SBK_SEPARATORPARAM_IDX                                       56
+#define SBK_INT3DTUPLE_IDX                                           25
+#define SBK_INT2DTUPLE_IDX                                           23
+#define SBK_PARAM_IDX                                                49
+#define SBK_GROUPPARAM_IDX                                           20
+#define SBK_SEPARATORPARAM_IDX                                       57
 #define SBK_BUTTONPARAM_IDX                                          6
 #define SBK_ANIMATEDPARAM_IDX                                        0
-#define SBK_INTPARAM_IDX                                             26
-#define SBK_INT2DPARAM_IDX                                           21
-#define SBK_INT3DPARAM_IDX                                           23
-#define SBK_STRINGPARAMBASE_IDX                                      60
-#define SBK_PATHPARAM_IDX                                            50
-#define SBK_OUTPUTFILEPARAM_IDX                                      46
-#define SBK_FILEPARAM_IDX                                            16
-#define SBK_STRINGPARAM_IDX                                          58
-#define SBK_STRINGPARAM_TYPEENUM_IDX                                 59
+#define SBK_STRINGPARAMBASE_IDX                                      61
+#define SBK_PATHPARAM_IDX                                            51
+#define SBK_OUTPUTFILEPARAM_IDX                                      47
+#define SBK_FILEPARAM_IDX                                            17
+#define SBK_STRINGPARAM_IDX                                          59
+#define SBK_STRINGPARAM_TYPEENUM_IDX                                 60
 #define SBK_BOOLEANPARAM_IDX                                         5
 #define SBK_CHOICEPARAM_IDX                                          7
 #define SBK_COLORPARAM_IDX                                           8
 #define SBK_DOUBLEPARAM_IDX                                          14
 #define SBK_DOUBLE2DPARAM_IDX                                        10
 #define SBK_DOUBLE3DPARAM_IDX                                        12
-#define SBK_PARAMETRICPARAM_IDX                                      49
-#define SBK_USERPARAMHOLDER_IDX                                      63
-#define SBK_IMAGELAYER_IDX                                           20
-#define SBK_TRACKER_IDX                                              62
-#define SBK_TRACK_IDX                                                61
-#define SBK_ROTO_IDX                                                 55
-#define SBK_ITEMBASE_IDX                                             27
+#define SBK_INTPARAM_IDX                                             27
+#define SBK_INT2DPARAM_IDX                                           22
+#define SBK_INT3DPARAM_IDX                                           24
+#define SBK_PARAMETRICPARAM_IDX                                      50
+#define SBK_PAGEPARAM_IDX                                            48
+#define SBK_USERPARAMHOLDER_IDX                                      64
+#define SBK_IMAGELAYER_IDX                                           21
+#define SBK_TRACKER_IDX                                              63
+#define SBK_TRACK_IDX                                                62
+#define SBK_ROTO_IDX                                                 56
+#define SBK_ITEMBASE_IDX                                             28
 #define SBK_BEZIERCURVE_IDX                                          3
-#define SBK_LAYER_IDX                                                28
-#define SBK_NODECREATIONPROPERTY_IDX                                 45
-#define SBK_STRINGNODECREATIONPROPERTY_IDX                           57
-#define SBK_FLOATNODECREATIONPROPERTY_IDX                            17
+#define SBK_LAYER_IDX                                                29
+#define SBK_NODECREATIONPROPERTY_IDX                                 46
+#define SBK_STRINGNODECREATIONPROPERTY_IDX                           58
+#define SBK_FLOATNODECREATIONPROPERTY_IDX                            18
 #define SBK_BOOLNODECREATIONPROPERTY_IDX                             4
-#define SBK_INTNODECREATIONPROPERTY_IDX                              25
+#define SBK_INTNODECREATIONPROPERTY_IDX                              26
 #define SBK_APPSETTINGS_IDX                                          2
-#define SBK_GROUP_IDX                                                18
-#define SBK_APP_IDX                                                  1
+#define SBK_GROUP_IDX                                                19
 #define SBK_EFFECT_IDX                                               15
-#define SBK_PYCOREAPPLICATION_IDX                                    51
-#define SBK_NatronEngine_IDX_COUNT                                   64
+#define SBK_APP_IDX                                                  1
+#define SBK_PYCOREAPPLICATION_IDX                                    52
+#define SBK_EXPRUTILS_IDX                                            16
+#define SBK_NatronEngine_IDX_COUNT                                   65
 
 // This variable stores all Python types exported by this module.
 extern PyTypeObject** SbkNatronEngineTypes;
@@ -142,10 +144,10 @@ extern SbkConverter** SbkNatronEngineTypeConverters;
 #define SBK_NATRONENGINE_STD_VECTOR_BOOL_IDX                         10 // const std::vector<bool > &
 #define SBK_NATRONENGINE_STD_VECTOR_INT_IDX                          11 // const std::vector<int > &
 #define SBK_NATRONENGINE_STD_LIST_EFFECTPTR_IDX                      12 // std::list<Effect * >
-#define SBK_NATRONENGINE_STD_MAP_QSTRING_NODECREATIONPROPERTYPTR_IDX 13 // const std::map<QString, NodeCreationProperty * > &
-#define SBK_NATRONENGINE_STD_LIST_QSTRING_IDX                        14 // std::list<QString >
-#define SBK_NATRONENGINE_STD_LIST_INT_IDX                            15 // const std::list<int > &
-#define SBK_NATRONENGINE_STD_MAP_IMAGELAYER_EFFECTPTR_IDX            16 // std::map<ImageLayer, Effect * >
+#define SBK_NATRONENGINE_STD_MAP_IMAGELAYER_EFFECTPTR_IDX            13 // std::map<ImageLayer, Effect * >
+#define SBK_NATRONENGINE_STD_MAP_QSTRING_NODECREATIONPROPERTYPTR_IDX 14 // const std::map<QString, NodeCreationProperty * > &
+#define SBK_NATRONENGINE_STD_LIST_QSTRING_IDX                        15 // std::list<QString >
+#define SBK_NATRONENGINE_STD_LIST_INT_IDX                            16 // const std::list<int > &
 #define SBK_NATRONENGINE_QLIST_QVARIANT_IDX                          17 // QList<QVariant >
 #define SBK_NATRONENGINE_QLIST_QSTRING_IDX                           18 // QList<QString >
 #define SBK_NATRONENGINE_QMAP_QSTRING_QVARIANT_IDX                   19 // QMap<QString, QVariant >
@@ -181,14 +183,10 @@ template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPAC
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::Int3DTuple >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_INT3DTUPLE_IDX]); }
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::Int2DTuple >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_INT2DTUPLE_IDX]); }
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::Param >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_PARAM_IDX]); }
-template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::PageParam >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_PAGEPARAM_IDX]); }
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::GroupParam >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_GROUPPARAM_IDX]); }
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::SeparatorParam >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_SEPARATORPARAM_IDX]); }
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::ButtonParam >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_BUTTONPARAM_IDX]); }
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::AnimatedParam >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_ANIMATEDPARAM_IDX]); }
-template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::IntParam >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_INTPARAM_IDX]); }
-template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::Int2DParam >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_INT2DPARAM_IDX]); }
-template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::Int3DParam >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_INT3DPARAM_IDX]); }
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::StringParamBase >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_STRINGPARAMBASE_IDX]); }
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::PathParam >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_PATHPARAM_IDX]); }
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::OutputFileParam >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_OUTPUTFILEPARAM_IDX]); }
@@ -201,7 +199,11 @@ template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPAC
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::DoubleParam >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_DOUBLEPARAM_IDX]); }
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::Double2DParam >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_DOUBLE2DPARAM_IDX]); }
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::Double3DParam >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_DOUBLE3DPARAM_IDX]); }
+template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::IntParam >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_INTPARAM_IDX]); }
+template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::Int2DParam >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_INT2DPARAM_IDX]); }
+template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::Int3DParam >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_INT3DPARAM_IDX]); }
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::ParametricParam >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_PARAMETRICPARAM_IDX]); }
+template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::PageParam >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_PAGEPARAM_IDX]); }
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::UserParamHolder >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_USERPARAMHOLDER_IDX]); }
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::ImageLayer >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_IMAGELAYER_IDX]); }
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::Tracker >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_TRACKER_IDX]); }
@@ -217,9 +219,10 @@ template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPAC
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::IntNodeCreationProperty >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_INTNODECREATIONPROPERTY_IDX]); }
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::AppSettings >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_APPSETTINGS_IDX]); }
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::Group >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_GROUP_IDX]); }
-template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::App >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_APP_IDX]); }
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::Effect >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_EFFECT_IDX]); }
+template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::App >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_APP_IDX]); }
 template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::PyCoreApplication >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_PYCOREAPPLICATION_IDX]); }
+template<> inline PyTypeObject* SbkType<NATRON_NAMESPACE::NATRON_PYTHON_NAMESPACE::ExprUtils >() { return reinterpret_cast<PyTypeObject*>(SbkNatronEngineTypes[SBK_EXPRUTILS_IDX]); }
 
 } // namespace Shiboken