+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## Version 2.1.3
 
+- Python: Add ExprUtils class that adds helpers for FBM and Perlin noise (taken from Walt Disney Animation SeExpr library)
 - Gui: sliders have a cleaner look with less ticks and a round handle 231c7f7
 - fix bug where the OFX plugin cache could be wrong if OpenGL was disabled 32c1532
 - fix dynamic kOfxSupportsTiles handling